Added Navigation Component
